added resourceActions grouping orders and evacuation tasks
