+++ _common/gameModel/libs/ClientScript/game/common/actions/radioActions.ts
@@ -1,3 +1,5 @@
+// EVALUATION_PRIORITY 20
+
 import {
   ActionTemplateUid,
   ActorId,
